Move middleware chain types into middleware::chain
