+++ gotham/src/middleware/mod.rs
@@ -6,6 +6,7 @@ use std::panic::RefUnwindSafe;
 use handler::HandlerFuture;
 use state::State;
 
+pub mod chain;
 pub mod session;
 
 /// `Middleware` has the opportunity to provide additional behaviour to the `Request` / `Response`
